feat(ai): rename agent-run to sandboxr
