+++ .antigravitycli/5917b7e7-f714-4fa8-8cf1-6a23b89a4e82.json
@@ -1,0 +1,1 @@
+/home/mkobit/.gemini/config/projects/5917b7e7-f714-4fa8-8cf1-6a23b89a4e82.json
