+++ src/chezmoi/.chezmoi.toml.tmpl
@@ -134,6 +134,9 @@ installation_method = "dotfiles.uv"
 installation_method = "dotfiles.uv"
 
 [data.local_bin_tools.agent_run]
+installation_method = "uninstall"
+
+[data.local_bin_tools.sandboxr]
 installation_method = "dotfiles.uv"
 
 # Fonts Installation Methods
