+++ src/chezmoi/.chezmoidata/starship.toml
@@ -3,11 +3,11 @@
 # `dig "starship" "custom_segments" dict .`
 
 [starship.custom_segments.agent_sandbox]
-# Show active sandbox profile when inside an agent-run sandbox.
-# AGENT_RUN_IN_SANDBOX=1 and AGENT_RUN_PROFILE are set by agent-run.
+# Show active sandbox profile when inside a sandboxr sandbox.
+# AGENT_RUN_IN_SANDBOX=1 and AGENT_RUN_PROFILE are set by sandboxr.
 format = '[$symbol$output]($style) '
 symbol = '⬡ '
 command = 'printf "%s" "${AGENT_RUN_PROFILE:-sandbox}"'
 when = 'test "$AGENT_RUN_IN_SANDBOX" = "1"'
 style = 'bold yellow'
-description = 'Active agent-run sandbox profile'
+description = 'Active sandboxr sandbox profile'
