+++ src/chezmoi/dot_config/ai-policy/claude/autonomous-settings.json.tmpl
@@ -1,4 +1,4 @@
-{{- /* Claude Code settings fragment for agent-run autonomous sandbox sessions,
+{{- /* Claude Code settings fragment for sandboxr autonomous sandbox sessions,
 passed via `claude --settings`. Prompts are bypassed because the bwrap sandbox
 plus credential absence is the boundary; nothing inside the sandbox is
 trusted, so per-tool deny lists are deliberately empty (they live in
