+++ src/chezmoi/dot_config/ai-policy/git/sandbox.gitconfig.tmpl
@@ -1,4 +1,4 @@
-# Global git config used inside agent-run sandboxes via GIT_CONFIG_GLOBAL.
+# Global git config used inside sandboxr sandboxes via GIT_CONFIG_GLOBAL.
 # Inherits the regular config, then forces signing off: ~/.gnupg is masked in
 # the sandbox, so signed commits would fail. Unsigned commits also mark
 # sandbox-made history for later human review.
