+++ src/chezmoi/dot_config/ai-policy/opencode/autonomous.json.tmpl
@@ -1,4 +1,4 @@
-{{- /* OpenCode config for agent-run autonomous sandbox sessions, passed via
+{{- /* OpenCode config for sandboxr autonomous sandbox sessions, passed via
 OPENCODE_CONFIG. Permission system fully open — the boundary is the bwrap
 sandbox plus credential absence, not anything settable inside the agent's
 config. */ -}}
