--- src/python/agent_sandbox/agent_sandbox/backend/__init__.py
@@ -1,5 +1,0 @@
-from agent_sandbox.backend.bwrap import BwrapBackend
-from agent_sandbox.backend.protocol import SandboxBackend, select_backend
-from agent_sandbox.backend.seatbelt import SeatbeltBackend
-
-__all__ = ["BwrapBackend", "SandboxBackend", "SeatbeltBackend", "select_backend"]
