--- src/python/agent_sandbox/agent_sandbox/profile/__init__.py
@@ -1,19 +1,0 @@
-from agent_sandbox.profile.loader import load_config, merge_cli_overrides, resolve_profile
-from agent_sandbox.profile.schema import (
-    Backend,
-    ConfigError,
-    Network,
-    Profile,
-    SandboxConfig,
-)
-
-__all__ = [
-    "Backend",
-    "ConfigError",
-    "Network",
-    "Profile",
-    "SandboxConfig",
-    "load_config",
-    "merge_cli_overrides",
-    "resolve_profile",
-]
