--- src/python/agent_sandbox/agent_sandbox/sandbox/__init__.py
@@ -1,4 +1,0 @@
-from agent_sandbox.sandbox.spec import SandboxSpec
-from agent_sandbox.sandbox.tool import adapt_command
-
-__all__ = ["SandboxSpec", "adapt_command"]
