+++ src/python/sandboxr/sandboxr/backend/__init__.py
@@ -1,0 +1,5 @@
+from sandboxr.backend.bwrap import BwrapBackend
+from sandboxr.backend.protocol import SandboxBackend, select_backend
+from sandboxr.backend.seatbelt import SeatbeltBackend
+
+__all__ = ["BwrapBackend", "SandboxBackend", "SeatbeltBackend", "select_backend"]
