+++ src/python/sandboxr/sandboxr/profile/__init__.py
@@ -1,0 +1,19 @@
+from sandboxr.profile.loader import load_config, merge_cli_overrides, resolve_profile
+from sandboxr.profile.schema import (
+    Backend,
+    ConfigError,
+    Network,
+    Profile,
+    SandboxConfig,
+)
+
+__all__ = [
+    "Backend",
+    "ConfigError",
+    "Network",
+    "Profile",
+    "SandboxConfig",
+    "load_config",
+    "merge_cli_overrides",
+    "resolve_profile",
+]
